shiftgen: Use exported error for unexpected row count
Summary:
The generated updater methods throw an error if the updated row count is not 1.
It would be nice for that error to be exported, so that a caller could react to that case.
It would also be nice to have the actual, incorrect row count added to the error as metadata. If the error was logged by the caller, for example, then we would be able to see what the actual row count was in application logs.

It was also noticed that there was an inconsistency in backtick quoting of column names. We are now quoting all column names in the generated code.

Reviewers: #backend, corver, andrewn, shiraaz, guy

Reviewed By: #backend, corver, andrewn, shiraaz, guy

Subscribers: guy, shiraaz, andrewn

Differential Revision: https://phabricator.corp.luno.com/D36007