shift: Add support for event metadata
Summary:
This adds support for optional event metadata to shift. This aims to provide for better audit functionality. Next step would be to write a `shift-console-audit` tool will create wellknown audit protobufs.

`shiftgen` doesn't generate metadata functions since it is orthogonal to sql boilerplate.

The interfaces `MetadataInserter` and `MetadataUpdater` is open enough to support dynamic metadata collection/generation. It is called right before event insertion with the arguments not present in the Inserter or Updater structs.

Reviewers: #backend, #fraud_team, neil

Reviewed By: #backend, neil

Subscribers: mdewit

Differential Revision: https://phabricator.corp.luno.com/D40043