shift: Add support for custom timestamps
Summary:
It is sometimes useful to have control over the timestamp columns that shift uses. We may choose not to store the created and updated times at all, deriving that information from the reflex event log. Or we may want custom behaviour - perhaps we need predictable time values for unit tests, or it is important that unmodified records have their created and updated times set to exactly equal values, or we just want non-shift column names.

Whatever the requirement, we can choose to turn off the default timestamp behaviour and set time values the same way we set values for other columns.

Reviewers: #pod-shield, #backend, kabrahams, corver

Reviewed By: #pod-shield, #backend, kabrahams, corver

Subscribers: corver, neil

Differential Revision: https://phabricator.corp.luno.com/D58155