Support string IDs in user tables
[external request](https://github.com/luno/shift/issues/4)

This is a change to support string primary keys in the user table that is used with shift.
This is accomplished by adding a generic implementation that allows int64 or string primary keys. We limit it to these two types (for now), since reflex supports EventsTable and EventsTableInt.

NOTE: The PR includes an example use case, but not the changes to update the code generation yet.

[gitlab issue](https://gitlab.com/lunomoney/product-engineering/pods/dx/work/-/issues/607)
[slack thread](https://luno.slack.com/archives/C034X4UFL/p1665384656626999)